Migrate tests to functional
