+++ discovery-service/src/test/java/org/zowe/apiml/discovery/functional/DiscoveryFunctionalTest.java
@@ -15,10 +15,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext
 public abstract class DiscoveryFunctionalTest {
 
+    protected static final String DISCOVERY_REALM = "API Mediation Discovery Service realm";
     @LocalServerPort
     protected int port;
